Move suggestion tests for the autocomplete package.
